+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- SwiftUI time for initial display and time for full display (#4596)
+
 ## 8.42.0-beta.2
 
 ### Fixes
@@ -25,7 +31,6 @@
 - Session replay transformed view masking (#4529)
 - Load integration from same binary (#4541)
 - Masking for fast animations #4574
-
 
 ### Improvements